// services/notification-service/src/notification/domain/model/notification-type.enum.ts
export enum NotificationType {
  ORDER_CREATED = 'ORDER_CREATED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  ORDER_SHIPPED = 'ORDER_SHIPPED',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  PAYMENT_COMPLETED = 'PAYMENT_COMPLETED',
  PAYMENT_REFUNDED = 'PAYMENT_REFUNDED',
  SELLER_NEW_ORDER = 'SELLER_NEW_ORDER',
  PRODUCT_APPROVED = 'PRODUCT_APPROVED',
  PRODUCT_REJECTED = 'PRODUCT_REJECTED',
  REVIEW_REPLIED = 'REVIEW_REPLIED',
  RETURN_REQUESTED = 'RETURN_REQUESTED',
  PAYOUT_COMPLETED = 'PAYOUT_COMPLETED',
  USER_REGISTERED = 'USER_REGISTERED',
  USER_PASSWORD_RESET = 'USER_PASSWORD_RESET',
}